State: move currencyCode out of currentUser to a separate reducer
