+++ client/my-sites/marketing/ultimate-traffic-guide/index.js
@@ -14,7 +14,8 @@ import { Button, CompactCard } from '@automattic/components';
 import QueryUserPurchases from 'calypso/components/data/query-user-purchases';
 import { getProductCost, isProductsListFetching } from 'calypso/state/products-list/selectors';
 import { isFetchingUserPurchases, getUserPurchases } from 'calypso/state/purchases/selectors';
-import { getCurrentUserCurrencyCode, getCurrentUserId } from 'calypso/state/current-user/selectors';
+import { getCurrentUserId } from 'calypso/state/current-user/selectors';
+import { getCurrentUserCurrencyCode } from 'calypso/state/currency-code/selectors';
 import { getSiteSlug } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
