+++ client/state/currency-code/selectors.js
@@ -1,0 +1,9 @@
+/**
+ * Returns the currency code for the current user.
+ *
+ * @param  {object}  state  Global state tree
+ * @returns {?string}        Current currency code
+ */
+export function getCurrentUserCurrencyCode( state ) {
+	return state.currencyCode;
+}
