+++ client/state/current-user/reducer.js
@@ -6,22 +6,9 @@ import { get, isEqual, reduce } from 'lodash';
 /**
  * Internal dependencies
  */
-import {
-	CURRENT_USER_RECEIVE,
-	SITE_RECEIVE,
-	SITE_PLANS_FETCH_COMPLETED,
-	SITES_RECEIVE,
-	PLANS_RECEIVE,
-	PRODUCTS_LIST_RECEIVE,
-} from 'calypso/state/action-types';
+import { CURRENT_USER_RECEIVE, SITE_RECEIVE, SITES_RECEIVE } from 'calypso/state/action-types';
 import { combineReducers, withSchemaValidation } from 'calypso/state/utils';
-import {
-	capabilitiesSchema,
-	currencyCodeSchema,
-	flagsSchema,
-	idSchema,
-	lasagnaSchema,
-} from './schema';
+import { capabilitiesSchema, flagsSchema, idSchema, lasagnaSchema } from './schema';
 import gravatarStatus from './gravatar-status/reducer';
 import emailVerification from './email-verification/reducer';
 
@@ -60,30 +47,6 @@ export const flags = withSchemaValidation( flagsSchema, ( state = [], action ) =
 	switch ( action.type ) {
 		case CURRENT_USER_RECEIVE:
 			return get( action.user, 'meta.data.flags.active_flags', [] );
-	}
-
-	return state;
-} );
-
-/**
- * Tracks the currency code of the current user
- *
- * @param  {object} state  Current state
- * @param  {object} action Action payload
- * @returns {object}        Updated state
- *
- */
-export const currencyCode = withSchemaValidation( currencyCodeSchema, ( state = null, action ) => {
-	switch ( action.type ) {
-		case PRODUCTS_LIST_RECEIVE: {
-			return Object.values( action.productsList )[ 0 ]?.currency_code ?? state;
-		}
-		case PLANS_RECEIVE: {
-			return get( action.plans, [ 0, 'currency_code' ], state );
-		}
-		case SITE_PLANS_FETCH_COMPLETED: {
-			return get( action.plans, [ 0, 'currencyCode' ], state );
-		}
 	}
 
 	return state;
@@ -137,7 +100,6 @@ export const lasagnaJwt = withSchemaValidation( lasagnaSchema, ( state = null, a
 export default combineReducers( {
 	id,
 	user,
-	currencyCode,
 	capabilities,
 	flags,
 	gravatarStatus,
