+++ client/state/current-user/schema.js
@@ -3,10 +3,6 @@ export const idSchema = {
 	minimum: 0,
 };
 
-export const currencyCodeSchema = {
-	type: [ 'string', 'null' ],
-};
-
 export const capabilitiesSchema = {
 	type: 'object',
 	additionalProperties: false,
