+++ client/state/current-user/selectors.js
@@ -101,16 +101,6 @@ export function getCurrentUserVisibleSiteCount( state ) {
 }
 
 /**
- * Returns the currency code for the current user.
- *
- * @param  {object}  state  Global state tree
- * @returns {?string}        Current currency code
- */
-export function getCurrentUserCurrencyCode( state ) {
-	return state.currentUser.currencyCode;
-}
-
-/**
  * Returns the date (of registration) for the current user.
  *
  * @param  {object}  state  Global state tree
