+++ client/state/current-user/test/reducer.js
@@ -7,15 +7,9 @@ import deepFreeze from 'deep-freeze';
 /**
  * Internal dependencies
  */
-import reducer, { id, capabilities, currencyCode } from '../reducer';
+import reducer, { id, capabilities } from '../reducer';
 import { serialize, deserialize } from 'calypso/state/utils';
-import {
-	CURRENT_USER_RECEIVE,
-	PLANS_RECEIVE,
-	SITE_RECEIVE,
-	SITE_PLANS_FETCH_COMPLETED,
-	SITES_RECEIVE,
-} from 'calypso/state/action-types';
+import { CURRENT_USER_RECEIVE, SITE_RECEIVE, SITES_RECEIVE } from 'calypso/state/action-types';
 import { useSandbox } from 'calypso/test-helpers/use-sinon';
 
 describe( 'reducer', () => {
@@ -27,7 +21,6 @@ describe( 'reducer', () => {
 		expect( reducer( undefined, {} ) ).to.have.keys( [
 			'id',
 			'user',
-			'currencyCode',
 			'capabilities',
 			'flags',
 			'gravatarStatus',
@@ -70,90 +63,6 @@ describe( 'reducer', () => {
 		test( 'will SERIALIZE current user', () => {
 			const state = serialize( id, 73705554 );
 			expect( state ).to.equal( 73705554 );
-		} );
-	} );
-
-	describe( '#currencyCode()', () => {
-		test( 'should default to null', () => {
-			const state = currencyCode( undefined, {} );
-			expect( state ).to.equal( null );
-		} );
-		test( 'should set currency code when plans are received', () => {
-			const state = currencyCode( undefined, {
-				type: PLANS_RECEIVE,
-				plans: [
-					{
-						product_id: 1001,
-						currency_code: 'USD',
-					},
-				],
-			} );
-			expect( state ).to.equal( 'USD' );
-		} );
-		test( 'should return current state when we have empty plans', () => {
-			const state = currencyCode( 'USD', {
-				type: PLANS_RECEIVE,
-				plans: [],
-			} );
-			expect( state ).to.equal( 'USD' );
-		} );
-		test( 'should update current state when we receive new plans', () => {
-			const state = currencyCode( 'USD', {
-				type: PLANS_RECEIVE,
-				plans: [
-					{
-						product_id: 1001,
-						currency_code: 'EUR',
-					},
-				],
-			} );
-			expect( state ).to.equal( 'EUR' );
-		} );
-		test( 'should return current state when we have empty site plans', () => {
-			const state = currencyCode( 'USD', {
-				type: SITE_PLANS_FETCH_COMPLETED,
-				plans: [],
-			} );
-			expect( state ).to.equal( 'USD' );
-		} );
-		test( 'should set currency code when site plans are received', () => {
-			const state = currencyCode( undefined, {
-				type: SITE_PLANS_FETCH_COMPLETED,
-				plans: [
-					{
-						productName: 'Free',
-						currencyCode: 'USD',
-					},
-				],
-			} );
-			expect( state ).to.equal( 'USD' );
-		} );
-		test( 'should update currency code when site plans are received', () => {
-			const state = currencyCode( 'USD', {
-				type: SITE_PLANS_FETCH_COMPLETED,
-				plans: [
-					{
-						productName: 'Free',
-						currencyCode: 'CAD',
-					},
-				],
-			} );
-			expect( state ).to.equal( 'CAD' );
-		} );
-		test( 'should persist state', () => {
-			const original = 'JPY';
-			const state = serialize( currencyCode, original );
-			expect( state ).to.equal( original );
-		} );
-		test( 'should restore valid persisted state', () => {
-			const original = 'JPY';
-			const state = deserialize( currencyCode, original );
-			expect( state ).to.equal( original );
-		} );
-		test( 'should ignore invalid persisted state', () => {
-			const original = 1234;
-			const state = deserialize( currencyCode, original );
-			expect( state ).to.equal( null );
 		} );
 	} );
