+++ client/state/reducer.js
@@ -15,6 +15,7 @@ import { reducer as httpData } from 'calypso/state/data-layer/http-data';
  * Reducers
  */
 import currentUser from './current-user/reducer';
+import currencyCode from './currency-code/reducer';
 import { reducer as dataRequests } from './data-layer/wpcom-http/utils';
 import documentHead from './document-head/reducer';
 import i18n from './i18n/reducer';
@@ -28,6 +29,7 @@ import userSettings from './user-settings/reducer';
 // Please do not add to this list. See #39261 and p4TIVU-9lM-p2 for more details.
 const reducers = {
 	currentUser,
+	currencyCode,
 	dataRequests,
 	documentHead,
 	httpData,
